backend/ml/requirements.txt
pymongo>=4.6.0
scikit-learn>=1.5.0
numpy>=1.26.0
joblib>=1.3.0
pandas>=2.1.0
python-dotenv>=1.0.0

# Ship Damage Detection Dependencies
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
flask>=2.3.0
flask-cors>=4.0.0
